finished task 4
