--- src/components/EditCourse/edit-course.css
@@ -1,110 +1,0 @@
-.top {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    margin-left: 20px;
-    margin-bottom: 20px;
-}
-
-.title-input {
-    display: flex;
-    flex-direction: column;
-}
-
-.title-input label{
-    margin-bottom: 5px;
-}
-
-.title-input input{
-    border: 1px solid orange;
-    height: 25px;
-    width: 500px;
-}
-
-.create-course-button {
-    margin-top: 25px;
-    margin-left: auto;
-    margin-right: 10px;
-}
-
-.create-course-button button{
-    border: 1px solid purple;
-    width: 120px;
-    height: 30px;
-}
-
-.textarea {
-    display: flex;
-    flex-direction: column;
-    margin-left: 20px;
-    margin-right: 20px;
-}
-
-.textarea textarea {
-    resize: none;
-    border: 1px solid #ffd500;
-    margin-top: 5px    
-}
-
-.bottom {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    border: 1px black solid;
-    margin: 20px;
-}
-
-.bottom p {
-    text-align: center;
-}
-
-.info {
-    display: flex;
-    flex-direction: column;
-}
-
-.author-info {
-    display: flex;
-    flex-direction: column;    
-    padding: 20px;
-}
-
-.author-info label {
-    margin-bottom: 5px;
-    margin-left: 0px;
-}
-
-.author-info input {
-    border: 1px solid orange;
-    margin-top: 0px;
-    width: 500px;
-    margin-bottom: 20px;
-}
-
-.author-info button {
-    border: 1px solid purple;
-    margin-left: 40%;
-    width: 120px;
-    height: 30px;
-}
-
-.add-remove-author {
-    display: flex;
-    flex-direction: column;
-    margin: 20px auto;
-}
-
-.author-list p{
-    text-align: left;    
-}
-
-.empty-list {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.author-info .duration {
-    font-size: 25px;
-    text-align: start;
-}
